Move IPC classes into individual files
