+++ src/logid/IPCServer.h
@@ -2,26 +2,11 @@
 #define LOGID_DBUSSERVER_H
 
 #include "ipc/dbus_server.h"
+#include "ipc/Root.h"
 
 #include <string>
 
 namespace logid {
-    namespace IPC
-    {
-        using namespace pizza::pixl;
-
-        class Root : public logiops_adaptor,
-                        public DBus::IntrospectableAdaptor,
-                        public DBus::ObjectAdaptor
-        {
-        public:
-            Root(DBus::Connection &connection): DBus::ObjectAdaptor(
-                    connection, "pizza.pixl.logiops") {}
-
-            virtual void Reload();
-        };
-    }
-
     class IPCServer
     {
     public:
